+++ src/pages/MemberForm/MemberForm.module.scss
@@ -1,4 +1,4 @@
-.wrap {
+body {
 	overflow-x: hidden;
 }
