+++ src/widgets/listPanels/MemberManagePanel/MemberManagePanel.module.scss
@@ -22,6 +22,7 @@
 		display: flex;
 		flex-direction: row-reverse;
 		align-self: flex-end;
+		align-items: center;
 	}
 }
 // thead {
